gesture module intergation

+++ vision/face_risk.py
@@ -1,3 +1,3 @@
 def compute_face_risk(cnn_probability, face_confidence):
     adjusted_risk = cnn_probability * face_confidence
-    return round(adjusted_risk, 2)
+    return float(round(adjusted_risk, 2))
